Add backup information to backup result func / cmd
Add more information about a completed backup. The current
implementation is a bit basic and probably need some improvement.
Currently we get the information for the latest backup (by running
`pgbackrest info` after the backup) and extract the information
needed by CloudNativePG. An other way to achieve that could be by
parsing the output of the command `pgbackrest backup` instead of
taking the backup and then getting the info of the latest backup
(which is a bit fragile).

Those information can be seen by inspecting the backup object by running
something similar to:
kubectl describe backups.postgresql.cnpg.io <backup-name>