Avoid default value when status code is missing
Now avoid using the default (zero) value when it's missing from
pgbackrest info whatever the reason. In such situation we will set that
field to nil.

We also split the function to check the repo to define if the stanza
exist into two parts. One function executes pgbackrest info and the
other one iterates over the result and determine if the stanza exist
(based on the status code) or not. This allow us to write some basic
test.